move some actions to plotActions global
