+++ app/assets/javascripts/components/plots/scatter_plot_container.js.jsx
@@ -1,50 +1,3 @@
-var addPlotData = function(plot_id, plot_json) {
-  return {
-    type: 'ADD_PLOT_DATA',
-    plot_id: plot_id,
-    series: plot_json.series,
-    mappings: plot_json.mappings,
-    data: plot_json.data
-  }
-}
-
-var cancelPlotConfig = function(plot_id) {
-  return {
-    type: 'CANCEL_PLOT_CONFIG',
-    plot_id: plot_id
-  }
-}
-
-var requestPlotData = function(plot, success, failure) {
-  var self = this;
-  var request = {
-    series: plot.requested_series,
-    mappings: plot.requested_mappings
-  };
-  return function(dispatch) {
-    $.ajax({
-      url: Routes.plot_json_path(),
-      method: 'POST',
-      data: JSON.stringify(request), 
-      dataType: 'json',
-      contentType: 'application/json',
-      success: function(plot_json) {
-        dispatch(addPlotData(plot.plot_id, plot_json))
-        success(plot_json);
-      },
-      error: function(error) {
-      }
-    })
-  }
-}
-
-var closePlot = function(plot_id) {
-  return {
-    type: 'CLOSE_PLOT',
-    plot_id: plot_id
-  }
-}
-
 ScatterPlotContainer = React.createClass({
   getInitialState: function() {
     return { mode: 'plot' }
